updating item model and connecting to seeding

+++ digital-monster/app/Http/Controllers/Admin/ItemController.php
@@ -9,19 +9,19 @@ use Illuminate\Support\Facades\Storage;
 
 class ItemController extends Controller
 {
-    protected $itemTypes = ['Case', 'Attack', 'Background', 'Consumable', 'Material'];
-    protected $rarityTypes = ['Free', 'Common', 'Uncommon', 'Rare', 'Legendary', 'Mythical'];
+    protected $itemTypes = ['Attack', 'Background', 'Consumable', 'Material'];
+    protected $rarityTypes = ['Common', 'Uncommon', 'Rare', 'Legendary'];
 
     public function index()
     {
         $items = Item::all()->groupBy('type');
         $icons = [
-            'Case' => 'fa-box',
             'Attack' => 'fa-bolt',
             'Background' => 'fa-image',
-            'Consumable' => 'fa-utensils',
-            'Material' => 'fa-cogs',
+            'Consumable' => 'fa-drumstick-bite',
+            'Material' => 'fa-boxes-stacked',
         ];
+
         return view('item.index', [
             'items' => $items,
             'rarityTypes' => $this->rarityTypes,
@@ -38,31 +38,67 @@ class ItemController extends Controller
 
     public function update(Request $request)
     {
+        $item = Item::findOrNew(session('item_id'));
+
         $validationRules = ([
-            'name' => 'required|string|max:255',
             'type' => 'required|string',
+            'rarity' => 'required|string',
+            'available' => 'required|numeric',
+            'name' => 'required|string|max:255',
             'price' => 'required|numeric|min:0',
             'max_quantity' => 'required|numeric|min:1',
-            'available' => 'required|numeric',
-            'rarity' => 'required|string',
         ]);
-        if (!session('item_id')) {
+
+        if (!$item->image) {
             $validationRules['image'] = 'required|image|mimes:png,jpg|max:2048';
         }
-        $request->validate($validationRules);
 
-        $itemData = $request->only(['name', 'type', 'effect', 'price', 'rarity', 'available', 'max_quantity']);
+        if ($request->input('type') == 'Background') {
+            if (!$item->image_1) {
+                $validationRules['image_1'] = 'required|image|mimes:png,jpg|max:2048';
+            }
+            if (!$item->image_2) {
+                $validationRules['image_2'] = 'required|image|mimes:png,jpg|max:2048';
+            }
+        }
+
+        if ($request->input('type') == 'Consumable') {
+            $validationRules['effect'] = 'required|string';
+        }
+
+        $validatedData = $request->validate($validationRules);
+        $itemData = $validatedData;
+
         if ($request->hasFile('image')) {
-            $path = $request->file('image')->store('items', 'public');
+            $path = $request->file('image')->store('item', 'public');
             $itemData['image'] = $path;
 
-            if (session('item_id')) {
-                $item = Item::findOrFail(session('item_id'));
+            if ($item->image) {
                 Storage::disk('public')->delete($item->image);
             }
         }
 
-        $item = session('item_id') ? Item::findOrFail(session('item_id'))->update($itemData) : Item::create($itemData);
+        if ($request->hasFile('image_1')) {
+            $path = $request->file('image_1')->store('item', 'public');
+            $itemData['image_1'] = $path;
+
+            if ($item->image_1) {
+                Storage::disk('public')->delete($item->image_1);
+            }
+        }
+
+        if ($request->hasFile('image_2')) {
+            $path = $request->file('image_2')->store('item', 'public');
+            $itemData['image_2'] = $path;
+
+            if ($item->image_2) {
+                Storage::disk('public')->delete($item->image_2);
+            }
+        }
+
+        $item->fill($itemData);
+        $item->save();
+
         $message = session('item_id') ? 'Item updated successfully.' : 'Item created successfully.';
 
         return redirect()->route('items.index')->with('success', $message);
@@ -71,7 +107,15 @@ class ItemController extends Controller
     public function destroy(Item $item)
     {
         $item = Item::findOrFail(session('item_id'));
-        Storage::disk('public')->delete($item->image);
+        if ($item->image) {
+            Storage::disk('public')->delete($item->image);
+        }
+        if ($item->image_1) {
+            Storage::disk('public')->delete($item->image_1);
+        }
+        if ($item->image_2) {
+            Storage::disk('public')->delete($item->image_2);
+        }
         $item->delete();
         return redirect()->route('items.index')->with('success', 'Item deleted successfully.');
     }
